extract functions
