+++ src/components/SlateEditor/plugins/SlateToolbar/SlateToolbar.jsx
@@ -13,14 +13,13 @@ import { Portal } from '../../../Portal';
 import ToolbarButton from './ToolbarButton';
 import { hasNodeOfType } from '../../utils';
 import { listTypes } from '../externalPlugins';
+import { handleClickBlock, handleClickMark, handleClickInline } from '../../utils/handleMenuClicks';
 
 const topicArticleElements = {
   mark: ['bold', 'italic', 'code', 'sub', 'sup'],
   block: ['quote', ...listTypes, 'heading-two', 'heading-three'],
   inline: ['link', 'mathml', 'concept'],
 };
-
-const DEFAULT_NODE = 'paragraph';
 
 const learningResourceElements = {
   mark: ['bold', 'italic', 'code', 'sub', 'sup'],
@@ -36,9 +35,6 @@ export const toolbarClasses = new BEMHelper({
 class SlateToolbar extends Component {
   constructor(props) {
     super(props);
-    this.onClickMark = this.onClickMark.bind(this);
-    this.onClickBlock = this.onClickBlock.bind(this);
-    this.onClickInline = this.onClickInline.bind(this);
     this.onButtonClick = this.onButtonClick.bind(this);
     this.portalRef = React.createRef();
     this.updateMenu = this.updateMenu.bind(this);
@@ -52,67 +48,13 @@ class SlateToolbar extends Component {
     this.updateMenu();
   }
 
-  onClickBlock(e, type) {
-    e.preventDefault();
-    const { editor } = this.props;
-    const { document, blocks } = editor.value;
-    const isActive = hasNodeOfType(editor, type);
-    if (type === 'quote') {
-      if (editor.isSelectionInBlockquote()) {
-        editor.unwrapBlockquote();
-      } else {
-        editor.wrapInBlockquote();
-      }
-    } else if (listTypes.includes(type)) {
-      const isListTypeActive = blocks.some(
-        block => !!document.getClosest(block.key, parent => parent.type === type),
-      );
-      // Current list type is active
-      if (isListTypeActive) {
-        editor.unwrapList();
-        // Current selection is list, but not the same type
-      } else if (editor.isSelectionInList()) {
-        editor.unwrapList();
-        editor.wrapInList(type);
-        // No list found, wrap in list type
-      } else {
-        editor.wrapInList(type);
-      }
-    } else {
-      editor.setBlocks(isActive ? DEFAULT_NODE : type);
-    }
-  }
-
-  onClickMark(e, type) {
-    e.preventDefault();
-    const { editor } = this.props;
-    editor.toggleMark(type);
-  }
-
-  onClickInline(e, type) {
-    e.preventDefault();
-    const { editor } = this.props;
-
-    if (type === 'footnote') {
-      editor
-        .moveToEnd()
-        .insertText('#')
-        .moveFocusForward(-1)
-        .wrapInline(type);
-    } else {
-      editor.withoutNormalizing(() => {
-        editor.wrapInline(type);
-      });
-    }
-  }
-
   onButtonClick(evt, kind, type) {
     if (kind === 'mark') {
-      this.onClickMark(evt, type);
+      handleClickMark(evt, this.props.editor, type);
     } else if (kind === 'block') {
-      this.onClickBlock(evt, type);
+      handleClickBlock(evt, this.props.editor, type);
     } else if (kind === 'inline') {
-      this.onClickInline(evt, type);
+      handleClickInline(evt, this.props.editor, type);
     }
   }
