+++ src/components/SlateEditor/utils/handleMenuClicks.js
@@ -1,0 +1,65 @@
+/**
+ * Copyright (c) 2021-present, NDLA.
+ *
+ * This source code is licensed under the GPLv3 license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+
+import { hasNodeOfType } from '.';
+import { listTypes } from '../plugins/externalPlugins';
+
+const DEFAULT_NODE = 'paragraph';
+
+function handleClickBlock(event, editor, type) {
+  event.preventDefault();
+  const { document, blocks } = editor.value;
+  const isActive = hasNodeOfType(editor, type);
+  if (type === 'quote') {
+    if (editor.isSelectionInBlockquote()) {
+      editor.unwrapBlockquote();
+    } else {
+      editor.wrapInBlockquote();
+    }
+  } else if (listTypes.includes(type)) {
+    const isListTypeActive = blocks.some(
+      block => !!document.getClosest(block.key, parent => parent.type === type),
+    );
+    // Current list type is active
+    if (isListTypeActive) {
+      editor.unwrapList();
+      // Current selection is list, but not the same type
+    } else if (editor.isSelectionInList()) {
+      editor.unwrapList();
+      editor.wrapInList(type);
+      // No list found, wrap in list type
+    } else {
+      editor.wrapInList(type);
+    }
+  } else {
+    editor.setBlocks(isActive ? DEFAULT_NODE : type);
+  }
+}
+
+function handleClickMark(event, editor, type) {
+  event.preventDefault();
+  editor.toggleMark(type);
+}
+
+function handleClickInline(event, editor, type) {
+  event.preventDefault();
+
+  if (type === 'footnote') {
+    editor
+      .moveToEnd()
+      .insertText('#')
+      .moveFocusForward(-1)
+      .wrapInline(type);
+  } else {
+    editor.withoutNormalizing(() => {
+      editor.wrapInline(type);
+    });
+  }
+}
+
+export { handleClickBlock, handleClickMark, handleClickInline };
